Move pol conv and stokes fields enum/bitflag types to top level
The concepts of polarization convention and identifying stokes cross
fields are more general that will [eventually] be used by other
concepts. Therefore, lift these (and rename appropriately) to the
top-level to be ready for that.

This also has the added benefit of making the type names shorter :-)
